refactor(toggleswitch): move ToggleSwitch to its folder with theme interface
#443 #234 #462
