+++ src/lib/components/FormControls/index.ts
@@ -1,2 +1,1 @@
 export * from './HelperText';
-export * from './ToggleSwitch';
