+++ src/lib/components/ToggleSwitch/ToggleSwitch.spec.tsx
@@ -2,10 +2,10 @@ import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import { FC, useState } from 'react';
 import { describe, expect, it, vi } from 'vitest';
-import { TextInput } from '../../TextInput';
+import { TextInput } from '../TextInput';
 import { ToggleSwitch } from './ToggleSwitch';
 
-describe('Components / Form controls / Toggle switch', () => {
+describe('Components / Toggle switch', () => {
   describe('A11y', () => {
     it('should have `role="switch"` by default', () => {
       render(<ToggleSwitch checked={false} label="Enable" onChange={console.log} />);
